git-svn-id: http://svn.code.sf.net/p/qtractor/code_attic/branches/osc@2411 304e7ffa-9037-4501-99c0-529757c9bbcc

+++ src/qtractorVstPlugin.cpp
@@ -55,8 +55,8 @@ typedef void (*XEventProc)(XEvent *);
 #endif
 
 #if !defined(VST_2_3_EXTENSIONS) 
-typedef long VstInt32;
-typedef long VstIntPtr;
+typedef int32_t  VstInt32;
+typedef intptr_t VstIntPtr;
 #define VSTCALLBACK
 #endif
